+++ packages/core/src/actions/providers/getProvider.ts
@@ -1,4 +1,4 @@
-import type { BaseProvider } from '@ethersproject/providers'
+import { providers } from 'ethers'
 
 import { getClient } from '../../client'
 
@@ -7,12 +7,13 @@ export type GetProviderArgs = {
   chainId?: number
 }
 
-export type GetProviderResult<TProvider extends BaseProvider = BaseProvider> =
-  TProvider
+export type GetProviderResult<
+  TProvider extends providers.BaseProvider = providers.BaseProvider,
+> = TProvider
 
-export function getProvider<TProvider extends BaseProvider = BaseProvider>({
-  chainId,
-}: GetProviderArgs = {}): GetProviderResult<TProvider> {
+export function getProvider<
+  TProvider extends providers.BaseProvider = providers.BaseProvider,
+>({ chainId }: GetProviderArgs = {}): GetProviderResult<TProvider> {
   const client = getClient<TProvider>()
   if (chainId && typeof client.config.provider === 'function')
     return client.config.provider({ chainId })