+++ packages/core/src/utils/getInjectedName.test.ts
@@ -3,13 +3,14 @@ import { getInjectedName } from './getInjectedName'
 describe.each`
   ethereum                                                             | expected
   ${undefined}                                                         | ${'Injected'}
-  ${{ isBraveWallet: true, isMetaMask: true }}                         | ${'Brave Wallet'}
-  ${{ isMetaMask: true }}                                              | ${'MetaMask'}
   ${{ isBraveWallet: true }}                                           | ${'Brave Wallet'}
+  ${{ isBraveWallet: true, isMetaMask: true }}                         | ${'Brave Wallet'}
   ${{ isCoinbaseWallet: true }}                                        | ${'Coinbase Wallet'}
   ${{ isFrame: true }}                                                 | ${'Frame'}
   ${{ isTally: true }}                                                 | ${'Tally'}
+  ${{ isTokenary: true, isMetaMask: true }}                            | ${'Tokenary'}
   ${{ isTrust: true }}                                                 | ${'Trust Wallet'}
+  ${{ isMetaMask: true }}                                              | ${'MetaMask'}
   ${{ providers: [{ isMetaMask: true }, { isCoinbaseWallet: true }] }} | ${['MetaMask', 'Coinbase Wallet']}
   ${{ providers: [{ isMetaMask: true }, { isFooWallet: true }, {}] }}  | ${['MetaMask', 'Unknown Wallet #1', 'Unknown Wallet #2']}
   ${{}}                                                                | ${'Injected'}