+++ packages/react/src/index.ts
@@ -30,6 +30,7 @@ export {
   useFeeData,
   useNetwork,
   useProvider,
+  useQuery,
   useSendTransaction,
   useSignMessage,
   useSignTypedData,